removed warnings

+++ src/core/networking/tcp/VRTCPClient.cpp
@@ -92,7 +92,7 @@ class TCPClient {
                 std::istream is(&buffer);
                 std::istreambuf_iterator<char> it(is);
                 copy_n( it, N-gN, std::back_inserter<std::string>(data) );
-                for (int i=0; i<=gN; i++) it++;
+                for (size_t i=0; i<=gN; i++) it++;
                 //data += "\n";
                 //cout << "  TCPClient, received: " << data.size() << " " << data << " " << int(data[0]) << endl;
                 if (onMessageCb) onMessageCb(data);
